Make Ref, Commit, Timestamp classes YAML representable
Also change the default constructor of Timestamp to take a time and
an offset and add a separate Timestamp.from_raw() method that parses
a raw string representation into a Timestamp object. This is because
pygit2 provides us with a (time, offset) tuple already so we don't
usually have to parse the data ourselves.

Adjust unit tests as necessary.