Improve code reuse in scenario tests, specify test store explicitly
There is now "scenario" Python module that can be used to load the
test output and parse it as YAML. The shell library now has a few
extra functions such as

    run_python <<-EOF
    print "Hello"
    EOF

and

    run_python_with_store <<-EOF
    print store.refs()
    EOF

that can be used to execute python code (the latter with a StorePool
and a Store preinitialised for the specified test store) in scenario
tests.

There is also

    run_python_test <<-EOF
    ...
    EOF

which uses the "scenario" module to preinitialise the variables
"output_raw" and "output_yaml" that can be accessed for checks.

All tests now have GIVEN statements of the form

    GIVEN a local standard test store "consonant-test-store-01"

rather than using aliases for the standard test stores like

    GIVEN a local simple kanban store

because the latter is not clear at all.