Implement LocalStore.objects(commit) with scenario tests
Unfortunately, we schemas don't play a role yet, so we cannot load
object properties into typed instances of subclasses of Property.
For now the values are taken as-is and are loaded into instances
of the Property base class rather than e.g. ReferenceProperty or
ListProperty.

The scenario tests for LocalStore.objects(commit) still need to be
completed, they currently only test a single commit.