Store objects in ObjectCache under "<UUID>,<SHA1>" instead of "<SHA1>"
The reason for this is that the SHA1 of properties.yaml is not enough
to uniquely identify objects, as different objects in different commits
and/or classes might share the same properties.yaml blob.

Caching via UUID + SHA1 makes sure that unpickling will result in
exactly the object we are expecting.