Add a basic implementation of applying update actions
Also rework the internals of consonant.store.local.LocalStore a bit
so that schemas, classes, objects etc. can be loaded from a Git tree
instead of requiring a commit object. This is important as actions
in transactions graduatelly create new trees to apply changes but,
apart from the final commit action, do not create new commits that
can then be accessed using the public Service API of LocalStore.

The commit also adds new scenario tests for updating an object as
well as creating and updating an object in separate actions of a
single transaction. This ensures that target object references
of both forms, another action or a UUID, are being tested.