Rename a few modules to make python-coverage-test-runner happy
There is a bug in python-coverage-test-runner that causes it to think
there is a lack of coverage when two .py modules have the same
basename. A fix was created a while ago but apparently that had
negative side effects, so this needs to be investigated further.

Short-term, renaming modules with duplicate names allows to run the
python-consonant test suite with an unmodified version of
python-coverage-test-runner, which is important.