Add a simple standalone Consonant web service based on Twisted
This simple service does not support authnz and app-specific hooks
yet. It does, however, support the complete basic web API as defined
in the Consonant specification. The transaction and query API are
yet to be implemented.

The service is based on Twisted's Site and Resource concepts but
does not make use of deferred request handling yet, which is
important for performance and scalability.