Add a simple standalone Consonant web service based on Twisted
This web service supports communication with YAML only at this
point. The complete basic and transaction web API is covered.
Not supported yet are JSON communication, authnz, app-specific
hooks and the query API.

The service is based on Twisted's Site and Resource concepts but
does not make use of deferred request handling yet, which is
important for performance and scalability.

In addition to the Consonant specification, the service handles
/commits/:sha1/... URLs to access content in particular commits
rather than just /refs/:ref/... which only allows retrieving
content for the latest commit in a ref. /commit/:sha1/... URLs
should probably be added to the Consonant specification.