WebService: Fix issues with loading from the latest commit in a ref
By passing a fixed commit in to all the pages, pages handling URLs
like "/refs/master/objects" would always fetch data from this fixed
commit passed in at construction time (when the server is started)
rather than dynamically picking the latest commit in the ref at the
time the requests come in.

This is now fixed.